// src/redux/tags/types.ts
export enum tagTypes {
  user = "user",
  services = "services",
  cart = "cart",
  booking = "booking",
  review = "review",
  feedback = "feedback",
  blog = "blog",
  faq = "faq",
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.services,
  tagTypes.cart,
  tagTypes.booking,
  tagTypes.review,
  tagTypes.feedback,
  tagTypes.blog,
  tagTypes.faq,
];